feat: Preenche lista com usuarios e atualiza ao excluir
