+++ auxia/src/components/list/ListItem.vue
@@ -1,4 +1,5 @@
 <template>
+  <Toast/>
   <div class="list-item">
     <div class="user-info">
       <span>{{ user.name }}</span>
@@ -30,6 +31,8 @@
 import { ref, computed, defineProps, defineEmits } from 'vue'
 import Dialog from 'primevue/dialog'
 import { useUserStore } from '@/stores/userStore'
+import { useToast, Toast } from 'primevue';
+
 
 interface User {
   name: string
@@ -40,7 +43,7 @@ interface User {
 
 const props = defineProps<{ user: User }>()
 const emit = defineEmits<{
-  (e: 'delete', payload: {success: boolean, email: string, error?: unknown }): void
+  (e: 'delete', user: User): void
   (e: 'edit', user: User): void
 }>()
 
@@ -48,13 +51,27 @@ const showModal = ref(false)
 
 const userStore = useUserStore();
 
+const toast = useToast();
+
 const confirmDelete = async () => {
   showModal.value = false
   try {
     await userStore.deleteByEmail(props.user.email)
-    emit('delete', { success: true, email: props.user.email })
+    toast.add({
+        severity: 'success',
+        summary: 'Sucesso!',
+        detail: 'Usuário excluido com sucesso!',
+        life: 3000
+      })
+    emit('delete', props.user)
   } catch (error) {
-    emit('delete', { success: false, email: props.user.email, error })
+    toast.add({
+      severity: 'error',
+      summary: 'Falha!',
+      detail: 'Falha ao excluir usuário!',
+      life: 3000
+    })
+    emit('delete', props.user)
   }
 }
